requirements.txt
numpy
pandas
pillow
PyYAML
matplotlib
tqdm
scikit-learn
torch
torchvision
tensorboard
